# backend/openapi.yaml
openapi: 3.1.0
info:
  title: "KAVACH \u2014 AI-Powered DevSecOps Platform"
  description: "KAVACH scans banking application repositories for vulnerabilities,\
    \ calculates a Banking Risk Score (BRS), maps findings to regulatory controls\
    \ (RBI / PCI-DSS / SWIFT), and generates AI-powered audit reports.\n\n## Pipeline\n\
    `GitHub webhook` \u2192 `Create Scan Job` \u2192 `Queue` \u2192 `Distributed workers`\
    \ \u2192 `Aggregation` \u2192 `Risk Engine` \u2192 `Compliance Engine` \u2192\
    \ `AI Explanation` \u2192 `Reports` \u2192 `Dashboard` \u2192 `Notifications`\
    \ \u2192 `Archive`\n\nA scan enters this pipeline either via `POST /scan/repository`\
    \ / `POST /scan` (an authenticated user submitting a URL or archive) or `POST\
    \ /webhooks/github` (an unauthenticated, HMAC-verified push event) \u2014 both\
    \ converge on the same queue \u2192 distributed-worker \u2192 aggregation path,\
    \ so everything downstream behaves identically regardless of how the scan was\
    \ triggered."
  contact:
    name: KAVACH Platform Team
    url: https://github.com/kavach-platform/kavach
  license:
    name: "Proprietary \u2014 All Rights Reserved"
  version: 1.0.0
paths:
...
  /api/v1/reports/{scan_job_id}:
    get:
      tags:
      - Reports
      summary: Get Report Status
      description: Check which reports are ready for download for a given scan job,
        and the status of any still in progress.
      operationId: get_report_status_api_v1_reports__scan_job_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: scan_job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Scan Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportPathsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/knowledge/upload:
    post:
      tags:
      - Knowledge Base
      summary: Upload Document
      description: "Upload a PDF/Markdown/text document. Returns immediately with\
        \ status\n\"pending\" \u2014 extraction/chunking/embedding runs asynchronously;\
        \ poll\nGET /knowledge/documents to see it advance to \"indexed\" (or \"failed\"\
        ).\n`tags` is a comma-separated string, e.g. \"pci-dss,payments\"."
      operationId: upload_document_api_v1_knowledge_upload_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_document_api_v1_knowledge_upload_post'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentUploadResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/knowledge/documents:
    get:
      tags:
      - Knowledge Base
      summary: List Documents
      operationId: list_documents_api_v1_knowledge_documents_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: document_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Document Type
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: tag
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Tag
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/knowledge/document/{document_id}:
    delete:
      tags:
      - Knowledge Base
      summary: Delete Document
      description: Deletes the document row (cascades to its chunks) and its on-disk
        file.
      operationId: delete_document_api_v1_knowledge_document__document_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Document Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDeleteResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/knowledge/search:
    post:
      tags:
      - Knowledge Base
      summary: Search Knowledge
      operationId: search_knowledge_api_v1_knowledge_search_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
...
  /api/v1/rag-operations/benchmark:
    post:
      tags:
      - RAG Operations
      summary: Run Benchmark
      description: "Triggers a live embed/search/rerank/(LLM) timing probe against\
        \ the\nknowledge base as it exists right now \u2014 see benchmark_service.py."
      operationId: run_benchmark_api_v1_rag_operations_benchmark_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BenchmarkResultSchema'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/rag-operations/search-analytics:
    get:
      tags:
      - RAG Operations
      summary: Get Search Analytics
      operationId: get_search_analytics_api_v1_rag_operations_search_analytics_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: feature
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Feature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchAnalyticsSummaryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/feedback:
    post:
      tags:
      - RAG Operations
      summary: Submit Feedback
      operationId: submit_feedback_api_v1_feedback_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackSubmitRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/feedback/summary:
    get:
      tags:
      - RAG Operations
      summary: Get Feedback Summary
      operationId: get_feedback_summary_api_v1_feedback_summary_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: feature
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Feature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackSummaryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    BenchmarkResultSchema:
      properties:
        ran_at:
          type: string
          title: Ran At
        stages:
          items:
            $ref: '#/components/schemas/BenchmarkStageSchema'
          type: array
          title: Stages
        total_duration_ms:
          type: number
          title: Total Duration Ms
        documents_indexed:
          type: integer
          title: Documents Indexed
        llm_configured:
          type: boolean
          title: Llm Configured
      type: object
      required:
      - ran_at
      - stages
      - total_duration_ms
      - documents_indexed
      - llm_configured
      title: BenchmarkResultSchema
    BenchmarkStageSchema:
      properties:
        stage:
          type: string
          title: Stage
        avg_duration_ms:
          type: number
          title: Avg Duration Ms
        detail:
          anyOf:
          - type: string
          - type: 'null'
          title: Detail
      type: object
      required:
      - stage
      - avg_duration_ms
      title: BenchmarkStageSchema
...
    Body_upload_document_api_v1_knowledge_upload_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
        version:
          anyOf:
          - type: string
          - type: 'null'
          title: Version
        author:
          anyOf:
          - type: string
          - type: 'null'
          title: Author
        tags:
          anyOf:
          - type: string
          - type: 'null'
          title: Tags
      type: object
      required:
      - file
      title: Body_upload_document_api_v1_knowledge_upload_post
...
    DocumentDeleteResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        filename:
          type: string
          title: Filename
        deleted:
          type: boolean
          title: Deleted
          default: true
      type: object
      required:
      - id
      - filename
      title: DocumentDeleteResponse
    DocumentListItem:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        filename:
          type: string
          title: Filename
        document_type:
          type: string
          title: Document Type
        version:
          type: string
          title: Version
        author:
          anyOf:
          - type: string
          - type: 'null'
          title: Author
        tags:
          items:
            type: string
          type: array
          title: Tags
        status:
          type: string
          title: Status
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
        file_size_bytes:
          type: integer
          title: File Size Bytes
        page_count:
          anyOf:
          - type: integer
          - type: 'null'
          title: Page Count
        chunk_count:
          type: integer
          title: Chunk Count
        uploaded_by_email:
          anyOf:
          - type: string
          - type: 'null'
          title: Uploaded By Email
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - filename
      - document_type
      - version
      - tags
      - status
      - file_size_bytes
      - chunk_count
      - created_at
      title: DocumentListItem
    DocumentListResponse:
      properties:
        total:
          type: integer
          title: Total
        documents:
          items:
            $ref: '#/components/schemas/DocumentListItem'
          type: array
          title: Documents
      type: object
      required:
      - total
      - documents
      title: DocumentListResponse
    DocumentUploadResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        filename:
          type: string
          title: Filename
        document_type:
          type: string
          title: Document Type
        version:
          type: string
          title: Version
        status:
          type: string
          title: Status
        tags:
          items:
            type: string
          type: array
          title: Tags
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - filename
      - document_type
      - version
      - status
      - tags
      - created_at
      title: DocumentUploadResponse
...
    FeedbackSubmitRequest:
      properties:
        feature:
          type: string
          maxLength: 32
          title: Feature
        reference_id:
          type: string
          maxLength: 255
          title: Reference Id
        rating:
          type: integer
          enum:
          - -1
          - 1
          title: Rating
        comment:
          anyOf:
          - type: string
            maxLength: 2000
          - type: 'null'
          title: Comment
      type: object
      required:
      - feature
      - reference_id
      - rating
      title: FeedbackSubmitRequest
    FeedbackSummaryResponse:
      properties:
        total_feedback:
          type: integer
          title: Total Feedback
        positive_count:
          type: integer
          title: Positive Count
        negative_count:
          type: integer
          title: Negative Count
        positive_rate:
          anyOf:
          - type: number
          - type: 'null'
          title: Positive Rate
      type: object
      required:
      - total_feedback
      - positive_count
      - negative_count
      title: FeedbackSummaryResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ReportPathsResponse:
      properties:
        scan_job_id:
          type: string
          format: uuid
          title: Scan Job Id
        pdf_available:
          type: boolean
          title: Pdf Available
        pdf_technical_available:
          type: boolean
          title: Pdf Technical Available
        sarif_available:
          type: boolean
          title: Sarif Available
        sbom_available:
          type: boolean
          title: Sbom Available
        unified_findings_available:
          type: boolean
          title: Unified Findings Available
        compliance_report_available:
          type: boolean
          title: Compliance Report Available
        csv_available:
          type: boolean
          title: Csv Available
        reports:
          items:
            $ref: '#/components/schemas/ReportStatusDetail'
          type: array
          title: Reports
          default: []
      type: object
      required:
      - scan_job_id
      - pdf_available
      - pdf_technical_available
      - sarif_available
      - sbom_available
      - unified_findings_available
      - compliance_report_available
      - csv_available
      title: ReportPathsResponse
    ReportStatusDetail:
      properties:
        report_type:
          type: string
          title: Report Type
        status:
          type: string
          title: Status
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
      type: object
      required:
      - report_type
      - status
      title: ReportStatusDetail
      description: "Per-report-type detail \u2014 lets a client show \"generating\"\
        /\"failed\" instead of just a missing boolean."
...
    SearchAnalyticsRecentEntry:
      properties:
        feature:
          type: string
          title: Feature
        query:
          type: string
          title: Query
        result_count:
          type: integer
          title: Result Count
        top_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Top Score
        latency_ms:
          type: number
          title: Latency Ms
        created_at:
          type: string
          title: Created At
      type: object
      required:
      - feature
      - query
      - result_count
      - latency_ms
      - created_at
      title: SearchAnalyticsRecentEntry
    SearchAnalyticsSummaryResponse:
      properties:
        total_searches:
          type: integer
          title: Total Searches
        average_latency_ms:
          anyOf:
          - type: number
          - type: 'null'
          title: Average Latency Ms
        average_result_count:
          anyOf:
          - type: number
          - type: 'null'
          title: Average Result Count
        zero_result_count:
          type: integer
          title: Zero Result Count
        zero_result_rate:
          anyOf:
          - type: number
          - type: 'null'
          title: Zero Result Rate
        recent_searches:
          items:
            $ref: '#/components/schemas/SearchAnalyticsRecentEntry'
          type: array
          title: Recent Searches
      type: object
      required:
      - total_searches
      - zero_result_count
      - recent_searches
      title: SearchAnalyticsSummaryResponse
    SearchRequest:
      properties:
        query:
          type: string
          maxLength: 2000
          minLength: 1
          title: Query
        top_k:
          type: integer
          maximum: 20.0
          minimum: 1.0
          title: Top K
          default: 5
        document_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Document Type
        tag:
          anyOf:
          - type: string
          - type: 'null'
          title: Tag
      type: object
      required:
      - query
      title: SearchRequest
    SearchResponse:
      properties:
        query:
          type: string
          title: Query
        took_ms:
          type: number
          title: Took Ms
        results:
          items:
            $ref: '#/components/schemas/SearchResultItem'
          type: array
          title: Results
      type: object
      required:
      - query
      - took_ms
      - results
      title: SearchResponse
    SearchResultItem:
      properties:
        document_id:
          type: string
          format: uuid
          title: Document Id
        filename:
          type: string
          title: Filename
        chunk_id:
          type: string
          format: uuid
          title: Chunk Id
        content:
          type: string
          title: Content
        similarity_score:
          type: number
          title: Similarity Score
        page_number:
          anyOf:
          - type: integer
          - type: 'null'
          title: Page Number
        heading:
          anyOf:
          - type: string
          - type: 'null'
          title: Heading
        section_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Section Path
      type: object
      required:
      - document_id
      - filename
      - chunk_id
      - content
      - similarity_score
      title: SearchResultItem
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
...